MOB-56 BeaconRegist Page
